@@ -1,51 +1,56 @@
 package com.ipartek.formacion.dgt.api;
+
+import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Set;
+
+import javax.validation.ConstraintViolation;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
+
 import org.apache.log4j.Logger;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ipartek.formacion.modelo.daos.MultaDAO;
-import com.ipartek.formacion.modelo.pojo.Multa;
-import com.ipartek.formacion.service.AgenteService;
-import com.ipartek.formacion.service.impl.AgenteServiceImpl;
+import com.ipartek.appMultas.modelo.dao.CocheDAO;
+import com.ipartek.appMultas.modelo.dao.MultaDAO;
+import com.ipartek.appMultas.modelo.pojo.Coche;
+import com.ipartek.appMultas.modelo.pojo.Multa;
+import com.mysql.jdbc.exceptions.jdbc4.MySQLIntegrityConstraintViolationException;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 
 @CrossOrigin
 @RestController
+@Api(tags = { "MULTAS" }, produces="application/json", value="Gestión de Multas")
+
 public class MultaController {
 
 	private final static Logger LOG = Logger.getLogger(MultaController.class);
 	public static MultaDAO multaDao;
-	private AgenteService agenteService;
-
+	private ValidatorFactory factory;
+	 private Validator validator;
+	
 	public MultaController() {
 		super();
 		multaDao = MultaDAO.getInstance();
-		agenteService = AgenteServiceImpl.getInstance();
-
+		//Crear Factoria y Validador
+		factory = Validation.buildDefaultValidatorFactory();
+		validator = factory.getValidator();
 	}
-
-
-	@RequestMapping(value = { "/api/agente/{id}/multas" }, method = RequestMethod.GET)
-	public ResponseEntity<ArrayList<Multa>> listar(@PathVariable long id) {
-		ResponseEntity<ArrayList<Multa>> response = new ResponseEntity<ArrayList<Multa>>(HttpStatus.NOT_FOUND);
-
-		try {
-
-			ArrayList<Multa> multas = new ArrayList<Multa>();
-			multas = (ArrayList<Multa>) agenteService.listarMultas(id);
-			
-			response = new ResponseEntity<ArrayList<Multa>>(multas, HttpStatus.OK);
-
-		} catch (Exception e) {
-			// TODO Auto-generated catch block
-			e.printStackTrace();
-		}
-
-		return response;
-	}
+	
+	
+	
+	
 }
